refactor(iceberg): split iceberg.go into focused files
